各Viewの中にベタ書きしていたViewを分離
