+++ StockWarChronicles/Domain/Tag.swift
@@ -30,21 +30,3 @@ final class Tag: Hashable {
         return .gray
     }
 }
-
-@Model
-final class CategoryTag {
-    @Attribute(.unique) var name: String
-    private var colorData: Data
-
-    init(name: String, color: Color) {
-        self.name = name
-        self.colorData = try! NSKeyedArchiver.archivedData(withRootObject: UIColor(color), requiringSecureCoding: false)
-    }
-
-    var color: Color {
-        if let uiColor = try? NSKeyedUnarchiver.unarchivedObject(ofClass: UIColor.self, from: colorData) {
-            return Color(uiColor)
-        }
-        return .gray
-    }
-}
